+++ src/backend/core/models.py
@@ -69,6 +69,7 @@ class ItemUploadStateChoices(models.TextChoices):
     """Defines the possible states of an item."""
 
     PENDING = "pending", _("Pending")
+    DUPLICATING = "duplicating", ("Duplicating")
     ANALYZING = "analyzing", _("Analyzing")
     SUSPICIOUS = "suspicious", _("Suspicious")
     FILE_TOO_LARGE_TO_ANALYZE = (
@@ -600,7 +601,11 @@ class Item(TreeModel, BaseModel):
                 }
             )
 
-        if self.created_at is None and self.type == ItemTypeChoices.FILE:
+        if (
+            self.created_at is None
+            and self.type == ItemTypeChoices.FILE
+            and self.upload_state != ItemUploadStateChoices.DUPLICATING
+        ):
             self.upload_state = ItemUploadStateChoices.PENDING
 
         if not self.path:
@@ -846,6 +851,12 @@ class Item(TreeModel, BaseModel):
             else (is_owner_or_admin or (user.is_authenticated and self.creator == user))
         )
         can_destroy = can_hard_delete and not is_deleted
+        can_duplicate = (
+            can_get
+            and user.is_authenticated
+            and self.type == ItemTypeChoices.FILE
+            and self.upload_state == ItemUploadStateChoices.READY
+        )
 
         return {
             "accesses_manage": is_owner_or_admin,
@@ -855,6 +866,7 @@ class Item(TreeModel, BaseModel):
             "children_create": can_create_children,
             "destroy": can_destroy,
             "download": can_get,
+            "duplicate": can_duplicate,
             "hard_delete": can_hard_delete,
             "favorite": can_get and user.is_authenticated,
             "link_configuration": is_owner_or_admin,