+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - ✨(frontend) add create folder and import file actions
 - 🐛(frontend) add action menu to mobile breadcrumbs
 - 🐛(frontend) app stabilization
+- ✨(backend) new action to duplicate a file item
 
 ### Changed